requirements.txt
openai
python-dotenv
google.generativeai
pathlib
chromadb
polars

langchain
langchain[openai]
langchain-chroma>=0.1.2
langchain-openai